+++ bpmn/MSOCommonBPMN/src/test/java/org/openecomp/mso/bpmn/common/PrepareUpdateAAIVfModuleTest.java
@@ -1,22 +1,22 @@
-/*- 
- * ============LICENSE_START======================================================= 
- * ONAP - SO 
- * ================================================================================ 
- * Copyright (C) 2017 AT&T Intellectual Property. All rights reserved. 
- * ================================================================================ 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
- * you may not use this file except in compliance with the License. 
- * You may obtain a copy of the License at 
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0 
- * 
- * Unless required by applicable law or agreed to in writing, software 
- * distributed under the License is distributed on an "AS IS" BASIS, 
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
- * See the License for the specific language governing permissions and 
- * limitations under the License. 
- * ============LICENSE_END========================================================= 
- */ 
+/*-
+ * ============LICENSE_START=======================================================
+ * ONAP - SO
+ * ================================================================================
+ * Copyright (C) 2017 AT&T Intellectual Property. All rights reserved.
+ * ================================================================================
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ============LICENSE_END=========================================================
+ */
 
 package org.openecomp.mso.bpmn.common;
 
@@ -45,37 +45,37 @@ public class PrepareUpdateAAIVfModuleTest extends WorkflowTest {
 	/**
 	 * Test the happy path through the flow.
 	 */
-	@Test	
+	@Test
 	@Deployment(resources = {
 			"subprocess/PrepareUpdateAAIVfModule.bpmn"
 		})
 	public void happyPath() throws IOException {
-		
+
 		logStart();
-		
-		String prepareUpdateAAIVfModuleRequest = FileUtil.readResourceFile("__files/VfModularity/PrepareUpdateAAIVfModuleRequest.xml"); 
-		
+
+		String prepareUpdateAAIVfModuleRequest = FileUtil.readResourceFile("__files/VfModularity/PrepareUpdateAAIVfModuleRequest.xml");
+
 		MockGetGenericVnfByIdWithDepth("skask", 1, "VfModularity/GenericVnf.xml");
 		MockPutGenericVnf("/skask/vf-modules/vf-module/supercool", "PCRF", 200);
 		MockPatchVfModuleId("skask", "supercool");
-		
+
 		String businessKey = UUID.randomUUID().toString();
 		Map<String, Object> variables = new HashMap<>();
 		variables.put("mso-request-id", "999-99-9999");
 		variables.put("isDebugLogEnabled","true");
 		variables.put("PrepareUpdateAAIVfModuleRequest", prepareUpdateAAIVfModuleRequest);
 		invokeSubProcess("PrepareUpdateAAIVfModule", businessKey, variables);
-		
+
 		Assert.assertTrue(isProcessEnded(businessKey));
 		String response = (String) getVariableFromHistory(businessKey, "PUAAIVfMod_updateVfModuleResponse");
 		Integer responseCode = (Integer) getVariableFromHistory(businessKey, "PUAAIVfMod_updateVfModuleResponseCode");
 		System.out.println("Subflow response code: " + responseCode);
 		System.out.println("Subflow response: " + response);
-		Assert.assertEquals(200, responseCode.intValue());
+		Assert.assertEquals(200, responseCode);
 		String heatStackId = (String) getVariableFromHistory(businessKey, "PUAAIVfMod_heatStackId");
 		System.out.println("Ouput heat-stack-id:" + heatStackId);
 		Assert.assertEquals("slowburn", heatStackId);
-		
+
 		logEnd();
 	}
 	
@@ -87,29 +87,29 @@ public class PrepareUpdateAAIVfModuleTest extends WorkflowTest {
 			"subprocess/PrepareUpdateAAIVfModule.bpmn"
 		})
 	public void badGet() throws IOException {
-		
+
 		logStart();
-		
-		String prepareUpdateAAIVfModuleRequest = FileUtil.readResourceFile("__files/VfModularity/PrepareUpdateAAIVfModuleRequest.xml"); 
+
+		String prepareUpdateAAIVfModuleRequest = FileUtil.readResourceFile("__files/VfModularity/PrepareUpdateAAIVfModuleRequest.xml");
 		MockGetGenericVnfById_404("skask[?]depth=1");
-		
+
 		String businessKey = UUID.randomUUID().toString();
 		Map<String, Object> variables = new HashMap<>();
 		variables.put("mso-request-id", "999-99-9999");
 		variables.put("isDebugLogEnabled","true");
 		variables.put("PrepareUpdateAAIVfModuleRequest", prepareUpdateAAIVfModuleRequest);
 		invokeSubProcess("PrepareUpdateAAIVfModule", businessKey, variables);
-		
+
 		Assert.assertTrue(isProcessEnded(businessKey));
 		String response = (String) getVariableFromHistory(businessKey, "PUAAIVfMod_getVnfResponse");
 		Integer responseCode = (Integer) getVariableFromHistory(businessKey, "PUAAIVfMod_getVnfResponseCode");
 		WorkflowException workflowException = (WorkflowException) getVariableFromHistory(businessKey, "WorkflowException");
 		System.out.println("Subflow response code: " + responseCode);
 		System.out.println("Subflow response: " + response);
-		Assert.assertEquals(404, responseCode.intValue());
+		Assert.assertEquals(404, responseCode);
 		Assert.assertNotNull(workflowException);
 		System.out.println("Subflow WorkflowException error message: " + workflowException.getErrorMessage());
-		
+
 		logEnd();
 	}
 	
@@ -121,25 +121,25 @@ public class PrepareUpdateAAIVfModuleTest extends WorkflowTest {
 			"subprocess/PrepareUpdateAAIVfModule.bpmn"
 		})
 	public void failValidation1() throws IOException {
-		
+
 		logStart();
-		
+
 		String prepareUpdateAAIVfModuleRequest = FileUtil.readResourceFile("__files/VfModularity/PrepareUpdateAAIVfModuleRequest.xml").replaceFirst("supercool", "lukewarm");
-		
+
 		MockGetGenericVnfByIdWithDepth("skask", 1, "VfModularity/GenericVnf.xml");
-		
+
 		String businessKey = UUID.randomUUID().toString();
 		Map<String, Object> variables = new HashMap<>();
 		variables.put("mso-request-id", "999-99-9999");
 		variables.put("isDebugLogEnabled","true");
 		variables.put("PrepareUpdateAAIVfModuleRequest", prepareUpdateAAIVfModuleRequest);
 		invokeSubProcess("PrepareUpdateAAIVfModule", businessKey, variables);
-		
+
 		WorkflowException workflowException = (WorkflowException) getVariableFromHistory(businessKey, "WorkflowException");
 		Assert.assertNotNull(workflowException);
 		System.out.println("Subflow WorkflowException error message: " + workflowException.getErrorMessage());
 		Assert.assertTrue(workflowException.getErrorMessage().startsWith("VF Module validation error: Orchestration"));
-		
+
 		logEnd();
 	}
 	
@@ -151,25 +151,25 @@ public class PrepareUpdateAAIVfModuleTest extends WorkflowTest {
 			"subprocess/PrepareUpdateAAIVfModule.bpmn"
 		})
 	public void failValidation2() throws IOException {
-		
+
 		logStart();
-		
+
 		String prepareUpdateAAIVfModuleRequest = FileUtil.readResourceFile("__files/VfModularity/PrepareUpdateAAIVfModuleRequest.xml").replaceFirst("supercool", "notsocool");
-		
+
 		MockGetGenericVnfByIdWithDepth("skask", 1, "VfModularity/GenericVnf.xml");		
-		
+
 		String businessKey = UUID.randomUUID().toString();
 		Map<String, Object> variables = new HashMap<>();
 		variables.put("mso-request-id", "999-99-9999");
 		variables.put("isDebugLogEnabled","true");
 		variables.put("PrepareUpdateAAIVfModuleRequest", prepareUpdateAAIVfModuleRequest);
 		invokeSubProcess("PrepareUpdateAAIVfModule", businessKey, variables);
-			
+
 		WorkflowException workflowException = (WorkflowException) getVariableFromHistory(businessKey, "WorkflowException");
 		Assert.assertNotNull(workflowException);
 		System.out.println("Subflow WorkflowException error message: " + workflowException.getErrorMessage());
 		Assert.assertTrue(workflowException.getErrorMessage().startsWith("VF Module validation error: VF Module"));
-		
+
 		logEnd();
 	}
 
@@ -181,32 +181,31 @@ public class PrepareUpdateAAIVfModuleTest extends WorkflowTest {
 			"subprocess/PrepareUpdateAAIVfModule.bpmn"
 		})
 	public void badPatch() throws IOException {
-		
+
 		logStart();
-		
-		String prepareUpdateAAIVfModuleRequest = FileUtil.readResourceFile("__files/VfModularity/PrepareUpdateAAIVfModuleRequest.xml"); 
-		
+
+		String prepareUpdateAAIVfModuleRequest = FileUtil.readResourceFile("__files/VfModularity/PrepareUpdateAAIVfModuleRequest.xml");
+
 		MockGetGenericVnfByIdWithDepth("skask", 1, "VfModularity/GenericVnf.xml");
 		MockAAIVfModuleBadPatch("/aai/v[0-9]+/network/generic-vnfs/generic-vnf/skask/vf-modules/vf-module/supercool", 404);
-		
+
 		String businessKey = UUID.randomUUID().toString();
 		Map<String, Object> variables = new HashMap<>();
 		variables.put("mso-request-id", "999-99-9999");
 		variables.put("isDebugLogEnabled","true");
 		variables.put("PrepareUpdateAAIVfModuleRequest", prepareUpdateAAIVfModuleRequest);
 		invokeSubProcess("PrepareUpdateAAIVfModule", businessKey, variables);
-		
+
 		Assert.assertTrue(isProcessEnded(businessKey));
 		String response = (String) getVariableFromHistory(businessKey, "PUAAIVfMod_updateVfModuleResponse");
 		Integer responseCode = (Integer) getVariableFromHistory(businessKey, "PUAAIVfMod_updateVfModuleResponseCode");
 		WorkflowException workflowException = (WorkflowException) getVariableFromHistory(businessKey, "WorkflowException");
 		System.out.println("Subflow response code: " + responseCode);
 		System.out.println("Subflow response: " + response);
-		Assert.assertEquals(404, responseCode.intValue());
+		Assert.assertEquals(404, responseCode);
 		Assert.assertNotNull(workflowException);
 		System.out.println("Subflow WorkflowException error message: " + workflowException.getErrorMessage());
-		
+
 		logEnd();
 	}
 }
-
